# .well-known/openapi.yaml
openapi: 3.1.0
info:
  title: Evolia Backend v4 (PRO)
  version: "4.0"
  description: >
    Full Evolia backend API:
    - Goal/Task management
    - Notion sync engine
    - AI Command system
    - Agents module
    - Health & system introspection

servers:
  - url: https://adnan-backend-v4.onrender.com

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:

    # -------------------------------
    # CORE MODELS
    # -------------------------------

    GoalModel:
      type: object
      required: [id, title, status, progress, created_at, updated_at]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        deadline: { type: string, nullable: true }
        parent_id: { type: string, nullable: true }
        priority: 
          type: string
          enum: [low, medium, high]
          nullable: true
        status: 
          type: string
          enum: [pending, in_progress, completed]
        progress:
          type: number
          minimum: 0
          maximum: 100
        children:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    TaskModel:
      type: object
      required: [id, title, status, created_at, updated_at]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        goal_id: { type: string, nullable: true }
        deadline: { type: string, nullable: true }
        priority:
          type: string
          enum: [low, medium, high]
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed]
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    AIResponse:
      type: object
      properties:
        success: { type: boolean }
        result: { nullable: true }
        message: { type: string, nullable: true }
        metadata: 
          type: object
          additionalProperties: true
        error: { type: string, nullable: true }

    AICommand:
      type: object
      required: [command]
      properties:
        command: { type: string }
        input: { nullable: true }
        agent: { type: string, nullable: true }
        params:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true

    AgentMessage:
      type: object
      required: [agent, content]
      properties:
        agent: { type: string }
        content: { type: string }
        type: { type: string, default: "message" }

    AgentProject:
      type: object
      required: [agent, title]
      properties:
        agent: { type: string }
        title: { type: string }
        description: { type: string }

    AgentStateUpdate:
      type: object
      required: [agent, state]
      properties:
        agent: { type: string }
        state: { type: string }

security:
  - ApiKeyAuth: []

paths:

  # -------------------------------
  # SYSTEM & HEALTH
  # -------------------------------

  /:
    get:
      summary: Root check
      responses:
        "200":
          description: Running status

  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Healthy

  /engine:
    get:
      summary: Master engine status
      responses:
        "200":
          description: Engine info

  /engine/state:
    get:
      summary: Get internal engine state
      responses:
        "200":
          description: State info

  /engine/progress:
    get:
      summary: Get progress engine status
      responses:
        "200":
          description: Progress info

  # -------------------------------
  # GOALS
  # -------------------------------

  /goals/all:
    get:
      summary: Get all goals
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: All goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalModel"

  /goals/create:
    post:
      summary: Create a new goal
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalModel"
      responses:
        "200":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalModel"

  /goals/{goal_id}:
    get:
      summary: Retrieve goal by ID
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: goal_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalModel"

    put:
      summary: Update goal
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: goal_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalModel"
      responses:
        "200":
          description: Updated goal

    delete:
      summary: Delete goal
      security:
        - ApiKeyAuth: []
      parameters:
        - name: goal_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted goal

  # -------------------------------
  # TASKS
  # -------------------------------

  /tasks/all:
    get:
      summary: Get all tasks
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: All tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskModel"

  /tasks/create:
    post:
      summary: Create a task
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskModel"
      responses:
        "200":
          description: Created task

  /tasks/{task_id}:
    get:
      summary: Get task by ID
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task

    put:
      summary: Update task
      security:
        - ApiKeyAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskModel"
      responses:
        "200":
          description: Updated task

    delete:
      summary: Delete task
      security:
        - ApiKeyAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted task

  # -------------------------------
  # SYNC
  # -------------------------------

  /sync/goals:
    post:
      summary: Sync goals with Notion (UP + DOWN)
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Sync result

  /sync/tasks:
    post:
      summary: Sync tasks with Notion (UP + DOWN)
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Sync result

  # -------------------------------
  # AI COMMANDS
  # -------------------------------

  /ai/command:
    post:
      summary: Execute AI command
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AICommand"
      responses:
        "200":
          description: AI command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIResponse"

  /ai/run:
    post:
      summary: Run high-level AI command
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AICommand"
      responses:
        "200":
          description: AI command result

  # -------------------------------
  # AGENTS
  # -------------------------------

  /agents/message:
    post:
      summary: Send message to agent exchange
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentMessage"
      responses:
        "200":
          description: Stored message

  /agents/messages:
    get:
      summary: Get recent messages
      security:
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of messages

  /agents/project:
    post:
      summary: Create agent project
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentProject"
      responses:
        "200":
          description: Project created

  /agents/state:
    post:
      summary: Update an agent's internal state
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentStateUpdate"
      responses:
        "200":
          description: Updated state